backend/requirements.txt
Django==3.2.3
djangorestframework==3.12.4
djoser==2.2.0
python-dotenv==1.0.0
regex==2023.8.8
psycopg2-binary==2.9.3
django-rest-auth==0.9.5
webcolors==1.13
pillow==10.0.1